token error

+++ src/hooks/useTaskManagement.ts
@@ -2,13 +2,23 @@
 
 import { useState, useEffect, useCallback } from 'react';
 import { Task } from '@/types';
+import { useSession } from 'next-auth/react';
 
 export const useTaskManagement = () => {
+
+  const { data: session } = useSession();
   const [tasks, setTasks] = useState<Task[]>([]);
 
   const fetchTasks = useCallback(async () => {
+    if (!session) return;
+
     try {
-      const response = await fetch('/api/tasks');
+      const response = await fetch('/api/tasks', {
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        credentials: 'include',
+      });
       if (response.ok) {
         const data = await response.json();
         setTasks(data);
@@ -18,18 +28,24 @@ export const useTaskManagement = () => {
     } catch (error) {
       console.error('Error fetching tasks:', error);
     }
-  }, []);
+  }, [session]);
 
   useEffect(() => {
-    fetchTasks();
-  }, [fetchTasks]);
+    if(session) {
+      fetchTasks();
+    }
+    
+  }, [session, fetchTasks]);
 
   const addTask = useCallback(async (task: Omit<Task, '_id'>) => {
+    if (!session) return;
+
     try {
       const response = await fetch('/api/tasks', {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify(task),
+        credentials: 'include',
       });
       if (response.ok) {
         const newTask = await response.json();
@@ -41,14 +57,17 @@ export const useTaskManagement = () => {
     } catch (error) {
       console.error('Error adding task:', error);
     }
-  }, [fetchTasks, tasks]);
+  }, [session]);
 
   const updateTask = useCallback(async (updatedTask: Task) => {
+    if (!session) return;
+
     try {
       const response = await fetch('/api/tasks', {
         method: 'PUT',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify(updatedTask),
+        credentials: 'include',
       });
       if (response.ok) {
         const updatedTaskData = await response.json();
@@ -60,14 +79,17 @@ export const useTaskManagement = () => {
     } catch (error) {
       console.error('Error updating task:', error);
     }
-  }, [fetchTasks, tasks]);
+  }, [session]);
   
   const deleteTask = useCallback(async (taskId: string) => {
+    if (!session) return;
+
     try {
       const response = await fetch('/api/tasks', {
         method: 'DELETE',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({ id: taskId }),
+        credentials: 'include'
       });
       if (response.ok) {
         setTasks(tasks.filter((task) => task._id !== taskId));
@@ -78,7 +100,7 @@ export const useTaskManagement = () => {
     } catch (error) {
       console.error('Error deleting task:', error);
     }
-  }, [fetchTasks, tasks]);
+  }, [session]);
 
-  return { tasks, addTask, updateTask, deleteTask };
+  return { tasks, addTask, updateTask, deleteTask, fetchTasks };
 };
